who_is_active
Fun stuff. This adds the additional memory counters to the main output instead of in the XML if you toggle the @advanced_meory_options=1.  The additional fields in the output are:
			requested_memory,
			granted_memory,
			required_memory,
			used_memory,
			max_used_memory,
			DOP,
			QueryCost

I left the used_memory mapped to the newer DMV used memory not (granted and process memused) . Note, I did not add the Deltas on all the new fields only left it on the used_memory, probably makes sense to add it to max_used_memory since it would vary between polls.