moved memory grants join; standardized grant output
I moved the memory grants join to a better place so that the DMV isn't hit more times than absolutely necessary. I also standardized the used_memory column output so that the format doesn't ever change, and I coalesced it with the old style of output so that we can always show some information here.