make all top-level counters page-specific
After playing with this a bit I thought it was confusing that the three new columns weren't on the same basis as the original used_memory column. So I decided to make them all consistent with the original one. If you want kb, you can get it from the XML.